requirements.txt
Flask~=1.1.2
python-dotenv~=0.15.0
Faker~=6.4.1
Flask-SQLAlchemy~=2.4.4